requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
anyio==3.7.1
asgiref==3.7.2
async-timeout==4.0.3
asynctest==0.13.0
attrs==23.1.0
bcrypt==4.0.1
boto==2.49.0
cached-property==1.5.2
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.3.0
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
cryptography==41.0.4
defusedxml==0.8.0rc2
distro==1.8.0
Django==3.2.22
django-filter==23.3
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==4.7.2
djoser==2.1.0
docker==6.1.3
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
drf-yasg==1.21.7
exceptiongroup==1.1.3
frozenlist==1.3.3
future==0.18.3
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.4
importlib-metadata==1.7.0
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jsonschema==3.2.0
MarkupSafe==2.1.3
multidict==6.0.4
oauthlib==3.2.2
packaging==23.2
paramiko==3.3.1
psycopg2==2.9.9
psycopg2-binary==2.8.6
pycparser==2.21
pycryptodome==3.19.0
PyJWT==2.8.0
PyNaCl==1.5.0
pyrsistent==0.19.3
python-dotenv==0.21.1
python3-openid==3.2.0
pytz==2023.3.post1
pywin32==306
PyYAML==5.4.1
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
sniffio==1.3.0
social-auth-app-django==4.0.0
social-auth-core==4.4.2
sqlparse==0.4.4
texttable==1.7.0
typing-extensions==4.7.1
uritemplate==4.1.1
urllib3==2.0.6
websocket-client==0.59.0
wrk==0.3.2
yarl==1.9.2
zipp==3.15.0
